+++ changelog.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Beta]
 
 ## [bt-1.4.1] - 2019-07-22
+##Fixes
+- Error with converter for artdata
+
+## [bt-1.4.1] - 2019-07-22
 ##Changes
 - Updated help command to reflect new commands in bt-1.4.0
 - Linked error handler for addallfolder and deleteallfolder